Helper superclass for SimpleAdapterBuilder
This exposes abstract methods and constructors that aids the building
of SimpleAdapter.  Constructor requires an Android Context, as well as
a List of classes that the specific builder builds on for SimpleAdapter.
the resource, from, and to are then exposed as abstract methods.

Subclasses of AbstractSimpleAdapterBuilder could now encapsulate all
steps necessary to build a SimpleAdapter for type T.

To use, given a sample class DataModel, a DataModelAdapterBuilder class
should be created, and all abstract methods and constructors implemented:

    public DataModelAdapterBuilder extends AbstactSimpleAdapterBuilder<DataModel> {
        public DataModelAdapter(final Context context, List<DataModel> data) {
            super(context, data);
        }

        protected List<? extends Map<String, ?>> data(List<DataModel> data) {
            ...
        }

        protected int resource() {
            ...
        }

        protected String[] from() {
            ...
        }

        protected int[] to() {
            ...
        }
    }

This could then be used in this manner:

    List<DataModel> data = new ArrayList<DataModel>();
    data.add(...);
    ...

    SimpleAdapter adapter = new DataModelAdapterBuilder(this, data).build();

Signed-off-by: Ealden Esto E. Escanan <ealden@teamcodeflux.com>