Define a custom configuration "provided" for dependencies that should be, in Maven's terms, provided
While provided = true works for Gradle, once a POM is generated they are treated
as compile.  We thus define a new configuration 'provided' to work around this.

What I noticed is that instead of having Maven dependencies in the provided
scope, the dependencies in question are not present at all.  I'm not sure if this
will cause problems, so let's commit this first and then test.

Signed-off-by: Ealden Esto E. Escanan <ealden@teamcodeflux.com>